fix: fix the delete constraints in intakes

+++ backend/src/modules/intakes/intakes.service.ts
@@ -123,10 +123,25 @@ export class IntakesService {
   }
 
   async remove(id: number) {
-    const deleteResult = await this.intakeRepository.delete(id);
+    const intake = await this.intakeRepository.findOne({
+      where: {
+        id,
+      },
+      relations: {
+        cohorts: true,
+      },
+    });
 
-    if (deleteResult.affected === 0) {
+    if (!intake) {
       throw new NotFoundException('Intake Not Found');
     }
+
+    if (intake.cohorts.length > 0) {
+      throw new BadRequestException(
+        'Cannot delete intake that contains cohorts',
+      );
+    }
+
+    await this.intakeRepository.delete(id);
   }
 }
